Issue13161: Adding invoice line
If we're writing on a new o2m, m2m or m2o record related to a new
record (for example an invoice line created altogether with the
invoice to which it belongs), ids contains a string. Calling
check_access_rule() in this condition will crash, so we just
blindly allow the operations.
From https://github.com/OCA/web/pull/710